refactor: extract backend podcast search mock handler to separate file
